backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

